refactor: Rework test-case as example
